Add basic fields and records
Fields are representations of SQL values. The Python descriptor protocol
is used to help ensure that the types are properly enforced.

Records represent a row in an SQL table. Use of a metaclass allows for
the use of __slots__ to reduce memory usage if many instances of a
record are created at once.