Add context-based values and RowEnumField
In typical database schemas, some of the columns hold values that link
the particular row to the wider context - for example a transaction id
or the row number of individual rows inserted under that transaction id.
These columns will typically be part of the primary key.

Various functions have been added that take a mutable dict parameter
context. This is used to retrieve the relevant context of the
transaction and (where necessary) to update it.

RowEnumField is a simple example of an SQLField subclass that uses the
context dict to track and increment a row-number. If an SQLRecord that
contains a RowEnumField has insert_sql called repeatedly with a
consistent dict provided for the context parameter, then the SQL insert
statement generated will have an increasing value for that RowEnumField
column.