RowEnumIntegerField and IDIntegerField are now set/gettable
Previously getting or setting these fields directly was forbidden as the
intention was that they should only be set in a particular context.
However this is very inflexible when it comes to working with data that
has been retrieved - in this case you do want to be able to see what row
or id. The get_context method on the fields will ignore any existing
values and produce new values from the context.

The SQLRecord get method was updated to take a context parameter. If
'None' then the existing values will be retrieved, otherwise new values
will be generated and the context may be updated. The values and
item_values methods also now take a context parameter and function
similarly.

Signed-off-by: James Humphry <git@binecho.co.uk>