SQLField now provides an SQL dialect-sensitive sql_repr method
This provides a way to cope with adaptors that may not deal well with
all of the standard Python types - for example the sqlite3 module will
not accept decimal.Decimal types unless they are converted to strings.

The alternative would be unconditional conversion to strings, but this
would be inefficient for the more flexible database adaptors.

Signed-off-by: James Humphry <git@binecho.co.uk>