Add SQLSequence class
This is a little messy as it extracts the SQL code for starting,
resetting and getting the next value from the dialect object passed in -
a form of dependency injection. However it is possible to create
subclasses that rely on database-specific methods and are faster.

For sqlite3, the default implementation creates a one-row table under
the name of the sequence.

Signed-off-by: James Humphry <git@binecho.co.uk>