Add SQLRecordList to hold multiple SQLRecords
SQLRecordList can be subclassed for a particular subclass of SQLRecord
to produce a class whose instances can hold ordered lists of records.
Fields on the underlying records are added to the SQLRecordList and
instances of the SQLRecordList will return generators giving the value
of that field on each underlying record in turn - so sum(recordlist.foo)
will give the sum of the foo fields of the records held. SQLRecordList
also implements many of the methods of the built in lists.

Signed-off-by: James Humphry <git@binecho.co.uk>