Remove some unnecessary checks for AttributeError
When SQLRecords are instantiated, all of the attributes defined by
SQLField descriptors and included in the class' __slots__ field are set
to None. There is therefore no need to check for AttributeError in the
SQLField.__get__ method as all instances will have a value set, even if
it is only None.

Signed-off-by: James Humphry <git@binecho.co.uk>