Change SequenceIntField to AbstractSequenceField
In order to make a field that represents a value from a sequence, it is
now necessary to subclass AbstractSequenceField with the sequence as a
class parameter, rather than an instance parameter. This is necessary so
the SQLTransaction subclasses can enforce that consistent sequences are
used for a given context.

Signed-off-by: James Humphry <git@binecho.co.uk>