Simplify SequenceIntField so it is only used in SQLTransaction
SequenceIntField needs only be a regular integer field with a link to
the SQLSequence from which it can be updated. When it is used as a
context field in an SQLTransaction, get_new_context will identify it and
update it from the database before including it in the context
context dictionary it returns.

In SQLRecord subclasses, a regular IDIntField can be used to extract
this value from the context dictionary. There is no need for
SequenceIntField to do both jobs.

Signed-off-by: James Humphry <git@binecho.co.uk>