SQLFields that return a value from the context should also store it
When get_context is used and an SQLField uses it in preference to
returning the value stored internally, it should store the value
returned.

This means that where RowEnumIntField is used, for example, the values
for that field may be initially left as None. After get_context has been
called (perhaps because the row has been inserted into the database) the
values will be updated with the row numbers actually used.

Signed-off-by: James Humphry <git@binecho.co.uk>