Allow better control and error-checking over NumericField quantization
NumericField now ensures that any values written to it have been
quantized in accordance with the scale and precision it was set up with.
Unless the inexact_quantize parameter is True, it will raise an error if
there is an attempt to write a value to it that cannot be exactly
represented with the quantization specified. It is also possible to
specify the rounding mode. Tests of the quantization behaviour have been
added.

Signed-off-by: James Humphry <git@binecho.co.uk>