Add QueryField to queries
This subclass of QueryField contains a reference to an SQLQuery. It is
intended for use as a context field inside an SQLTransaction. When
get_new_context is called on the SQLTransaction, the SQLQuery inside
the QueryField will be executed using the SQLTransaction's context
dictionary as it exists at that point and the field will be set to the
first value returned from the query (the query has to return a single
column).

This has to be in pyxact.queries rather than pyxact.fields to prevent
circular dependency problems.

Signed-off-by: James Humphry <git@binecho.co.uk>