ColumnsConstraint now takes Python column names rather than SQL names
This makes SQLRecordMetaClass a bit uglier as it has to go through and
amend each ColumnsConstraint to add an sql_column_names list of the SQL
names of the columns referenced by column_names attribute.

SQLRecordMetaClass also calculates whether the ColumnsConstraint refers
to a superkey (all non-null, unique columns).

ForeignKeyConstraint is made a subclass of ColumnsConstraint to avoid
duplicate work.

Signed-off-by: James Humphry <git@binecho.co.uk>