get_context() and get_new_context() now only return non-None values
This is for both SQLRecord and SQLTransaction and will make it easier
to implement more complex usage patterns - i.e. using a transaction ID
on its own to context_select a record from a database and then filling
out the rest of the context from the records with
set_context_from_records. Without this change, it will be hard to avoid
having comparisons to NULL/None in SQL statements.

Signed-off-by: James Humphry <git@binecho.co.uk>