Split out SQLRecord and SQLTable (along with metaclasses)
SQLRecord now describes the basic record class with SQLField attributes
but it has no methods relating to SQL operations, does nothing with
SQLConstraints and lacks information about any SQL table or schema. It
is the base for SQLTable and (in the future) SQLView classes, and may
be used directly as the return type from SQLQueries.

SQLTable processes SQLConstraint attributes, contains SQL operation
methods and also has additional attributes that link the class to an
SQL table.

Signed-off-by: James Humphry <git@binecho.co.uk>