Allow TimestampField to optionally take values from the context dict
If the context_used parameter is set, the name specified will be
searched for in the context dictionary (if provided) and the value in
the dictionary used in preference to the value in the SQLRecord
instance.

This might be preferable behaviour for all SQLField, rather than having
separate context-aware and context-unaware types.

Signed-off-by: James Humphry <git@binecho.co.uk>