Make all SQLField context-sensitive if context_used param is set
If any SQLField is created with the context_used parameter set to a
non-None value, this field will now look for a context value under that
name when get_context() is called. It will set itself to the value in
the context dictionary, doing any type checking and conversion in the
process. It will finally return that value rather than the value stored
in the instance of SQLRecord.

This is simpler and more consistent than having duplicate subclasses of
SQLField, some of which are context aware and some of which aren't. This
change allows ContextIntField to be eliminated.

Signed-off-by: James Humphry <git@binecho.co.uk>