Prefix all SQLTransaction methods with _
In order to minimise name clashes between user-defined attributes and
methods defined on SQLTransaction, all of the methods now have the
prefix _.

The Python convention is that this makes them private, which is
inconsistent with their public nature. However this solution to the
name clash issue is the one adopted by collections.namedtuple, so it has
some precedent.

Signed-off-by: James Humphry <git@binecho.co.uk>