Make SQLField take a query param to use in the default update method
The 'query' param of SQLField should indicate a subclass of SQLQuery. If
present, the default update method of SQLField will run that query
expecting a single value result that the field can be set to. This is
simpler than having lots of query variants of fields, or some sort of
complicated decorator.

Signed-off-by: James Humphry <git@binecho.co.uk>