Allow variable transaction isolation levels to be used
A new enum, IsolationLevel, has five settings that correspond to the
four levels of transaction isolation defined in the SQL standards, plus
one meta-level that indicates that pyxact should not attempt to manage
transactions itself.

SQLDialect now has three new methods that begin, commit and rollback
transactions. An isolation_level parameter indicates what transaction
isolation is required. If MANUAL_TRANSACTIONS is used, the methods will
do nothing. The begin method returns a context manager so can be used
to create a 'with' block that starts a new transaction that will be
committed at the end (or rolled-back if an exception occurs).

SQLTransaction takes a new isolation_level class parameter. This
isolation level is used for all transactions initiated by methods on the
class. The methods now all roll-back the transaction if an exception
occurs.

Signed-off-by: James Humphry <git@binecho.co.uk>

Signed-off-by: James Humphry <git@binecho.co.uk>