Create separate refresh and update methods on SQLField
The refresh method may run queries to fetch data from the database, but
will not increment sequences, update timestamps or do other operations
that might modify the database or generate 'new' data. The default
implementation gets data by executing the query attached to the SQLField
(if any) and setting the field to the result.

The update method is allow to make modifications, but by default it just
calls the refresh method.

Signed-off-by: James Humphry <git@binecho.co.uk>