Change SQLTransaction._get_context to call SQLField.get_context
Now, like the _get_refreshed_context and _get_updated_context it will
form a context dictionary and pass it to the underlying SQLFields'
get_context methods. While fields attached to an SQLTransaction will
probably not actually want a context dictionary (they should be
providing one, not consuming one), this is more consistent with the
naming of the method, and the way things work in SQLRecord.

Signed-off-by: James Humphry <git@binecho.co.uk>