Enhance ForeignKeyConstraint to allow for MATCH, ON DELETE, DEFERRABLE
In order to allow proper updates and deletes on SQLTransaction which
contain internal foreign key linkages, it is necessary to enhance the
support for foreign key constraints. By default ON DELETE and ON UPDATE
actions are NO ACTION (so constraint-breaking changes will cause errors)
but DEFERRABLE INITIALLY DEFERRED is also the default, so the
constraints will only be checked at the end of the transaction. This
means that SQLTransaction does not have to consider the dependency tree
to carry out the operations in a particular order to avoid premature
constraint violation errors.

Signed-off-by: James Humphry <git@binecho.co.uk>