Initialise all SQLRecordList fields in SQLTransaction with an empty list
When an SQLTransaction subclass is instantiated without data being
provided for all attributes, it is reasonable to leave SQLField and
SQLTable attributes as None. However the default initial state of an
SQLRecordList attribute should be an empty SQLRecordList, because no
question arises about having to guess suitable values for an empty list.

Signed-off-by: James Humphry <git@binecho.co.uk>