SQLTransaction updates and deletions should be done in reverse order
If SQLTransaction are written so that SQLTable included later in the
definition depend (via foreign key constraints) on those defined earlier
in the definition, but not vice-versa, then it makes sense to do
deletions and updates in reverse order. Of course a real solution is to
set DEFERRABLE_INITIALLY_DEFERRED on the constraint, but this might not
be practical in some cases where triggers are being used.

Signed-off-by: James Humphry <git@binecho.co.uk>