Add basic ltree support - pyxact.postgresql.ltree
This subclasses the builtin Python string type to create a special case
that represents labels of data stored in a hierarchical tree-like
structure. These map to the 'ltree' type in the 'ltree' extension, a
standard extension module provided with PostgreSQL. A fall-back to
storage as TEXT is provided for other databases, but obviously the rich
set of operators provided by the PostgreSQL extension will not be
available.

Signed-off-by: James Humphry <git@binecho.co.uk>