Change the way TransactionContext begins transactions
When the begin_transaction class method is called on a dialect object,
the actual execution of the SQL to start a transaction is now deferred
to the __enter__ method of the TransactionContext returned. This means
it cannot be used outside of a context manager (i.e. a 'with' statement)
but it is unlikely to be used like that anyway.

This means that the TransactionContext can be updated to be an async
context manager at the same time.

Signed-off-by: James Humphry <git@binecho.co.uk>