Remove dialect parameter from API
In practice, it is highly unlikely that mixed database dialects will be
used in any program. Having dialect parameters everywhere was confusing
and error-prone. It is simpler to have the user of pyxact set
dialects.DefaultDialect at the start of the program, and to let every
other part of the program use that value if necessary.

This is a balancing act where I think YAGNI outweighs the 'code smell'
of using what is effectively a global variable - mainly because the
global variable will not actually vary during any program execution.

Signed-off-by: James Humphry <git@binecho.co.uk>

Signed-off-by: James Humphry <git@binecho.co.uk>

Signed-off-by: James Humphry <git@binecho.co.uk>